OddDataWins to float

+++ utils/structs.go
@@ -193,8 +193,8 @@ type PickStruct struct {
 }
 type PickDetails struct {
 	Picks           []Picks   `json:"picks"`
-	OddDataUnits    float64       `json:"oddDataUnits"`
-	OddDataWins     int       `json:"oddDataWins"`
+	OddDataUnits    float64   `json:"oddDataUnits"`
+	OddDataWins     float64   `json:"oddDataWins"`
 	PickId          string    `json:"pickId"`
 	Gamestatus      []string  `json:"gamestatus"`
 	PickStatus      string    `json:"pickStatus"`
